requirements (1).txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
crispy==0.7.4
crispy-bootstrap5==2024.2
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.0
Django==4.2.11
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-jinja==2.11.0
fabio==2023.10.0
fonttools==4.50.0
frozenlist==1.4.1
h11==0.14.0
h5py==3.10.0
hdf5plugin==4.4.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.3.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pillow==10.2.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.1
redis==5.0.2
service-identity==24.1.0
silx==2.0.1
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.2
djangorestframework